AtlasAI/Atlas_SwissAgent/usable_snippets/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swissagent"
version = "0.2.0"
description = "Offline AI-powered development platform with Copilot-style IDE, Ollama/Open WebUI support, and 195+ tools"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "toml>=0.10.2",
    "gitpython>=3.1.41",
    "Pillow>=10.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "anyio>=4.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
swissagent = "core.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
